refactor: move RelativeURL() function to another package (to avoid import cycle later)
