+++ eval/lib/url.go
@@ -1,18 +1,16 @@
 package lib
 
 import (
-	"fmt"
 	"net/url"
-	"regexp"
 	"strings"
 
 	"github.com/zclconf/go-cty/cty"
 	"github.com/zclconf/go-cty/cty/function"
+
+	url1 "github.com/avenga/couper/eval/lib/url"
 )
 
 var (
-	// https://datatracker.ietf.org/doc/html/rfc3986#page-50
-	regexParseURL   = regexp.MustCompile(`^(([^:/?#]+):)?(//([^/?#]*))?([^?#]*)(\?([^#]*))?(#(.*))?`)
 	UrlEncodeFunc   = newUrlEncodeFunction()
 	RelativeUrlFunc = newRelativeUrlFunction()
 )
@@ -53,7 +51,7 @@ func newRelativeUrlFunction() function.Function {
 		Type: function.StaticReturnType(cty.String),
 		Impl: func(args []cty.Value, _ cty.Type) (ret cty.Value, err error) {
 			absURL := strings.TrimSpace(args[0].AsString())
-			relURL, err := RelativeURL(absURL)
+			relURL, err := url1.RelativeURL(absURL)
 			if err != nil {
 				return cty.StringVal(""), err
 			}
@@ -61,25 +59,4 @@ func newRelativeUrlFunction() function.Function {
 			return cty.StringVal(relURL), nil
 		},
 	})
-}
-
-func RelativeURL(absURL string) (string, error) {
-	if !strings.HasPrefix(absURL, "/") && !strings.HasPrefix(absURL, "http://") && !strings.HasPrefix(absURL, "https://") {
-		return "", fmt.Errorf("invalid url given: %q", absURL)
-	}
-
-	// Do not use the result of url.Parse() to preserve the # character in an emtpy fragment.
-	if _, err := url.Parse(absURL); err != nil {
-		return "", err
-	}
-
-	// The regexParseURL garanties the len of 10 in the result.
-	urlParts := regexParseURL.FindStringSubmatch(absURL)
-
-	// The path must begin w/ a slash.
-	if !strings.HasPrefix(urlParts[5], "/") {
-		urlParts[5] = "/" + urlParts[5]
-	}
-
-	return urlParts[5] + urlParts[6] + urlParts[8], nil
 }
